Remove empty apps: behavior, electrophys, equipment, imaging
