+++ alyx/backup.sh
@@ -1,0 +1,1 @@
+pg_dump -cOx -U cyrille -h localhost alyx -f alyx_full.local.sql
